+++ src/mflux/models/z_image/cli/z_image_turbo_generate.py
@@ -58,7 +58,7 @@ def main():
                 negative_prompt=args.negative_prompt,
             )
             # 4. Save the image
-            image.save(path=args.output.format(seed=seed), export_json_metadata=args.metadata)
+            image.save(path=args.output.format(seed=seed), export_json_metadata=args.metadata, overwrite=args.replace)
     except (StopImageGenerationException, PromptFileReadError) as exc:
         print(exc)
     finally: